refactor(): Change structure; Migrate to go errors 113
