+++ internal/processor/parser/report1.go
@@ -3,12 +3,11 @@ package parser
 
 import (
 	"bufio"
+	"fmt"
 	"io"
 	"regexp"
 
-	"github.com/pkg/errors"
-
-	"github.com/oleg-balunenko/spamassassin-parser/internal/models"
+	"github.com/oleg-balunenko/spamassassin-parser/internal/processor/models"
 )
 
 var (
@@ -42,15 +41,15 @@ func (rp report1Parser) Parse(data io.Reader) (models.Report, error) {
 
 		matches := reType1.FindStringSubmatch(line)
 		if len(matches) == 0 {
-			return emptyReport, errors.Errorf("failed to find matches for regex [line num: %d], [line: %s]",
+			return emptyReport, fmt.Errorf("failed to find matches for regex [line num: %d], [line: %s]",
 				lnum, line)
 		}
 
 		if matches[colScore] != "" {
 			h, err := makeHeader(matches[colScore], matches[colTag], matches[colDescr])
 			if err != nil {
-				return emptyReport, errors.Wrapf(err,
-					"failed to make header [line num: %d], [line: %s]", lnum, line)
+				return emptyReport, fmt.Errorf(
+					"failed to make header [line num: %d], [line: %s]: %w", lnum, line, err)
 			}
 
 			score += h.Score
