+++ internal/processor/parser/reporttype_string.go
