+++ internal/processor/testdata/report1.golden.json
