+++ internal/processor/testdata/report2.golden.json
