+++ pkg/utils/utils.go
@@ -4,16 +4,13 @@ package utils
 import (
 	"bytes"
 	"encoding/json"
+	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
 
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
-
-	"github.com/oleg-balunenko/spamassassin-parser/internal/models"
 )
 
 // PrettyPrint appends to passed struct indents and returns a human readable form of struct.
@@ -21,16 +18,16 @@ import (
 func PrettyPrint(v interface{}, prefix string, indent string) (string, error) {
 	b, err := json.Marshal(v)
 	if err != nil {
-		return "", errors.Wrap(err, "failed to marshal")
+		return "", fmt.Errorf("failed to marshal: %w", err)
 	}
 
 	var out bytes.Buffer
 	if err := json.Indent(&out, b, prefix, indent); err != nil {
-		return "", errors.Wrap(err, "failed to indent")
+		return "", fmt.Errorf("failed to indent: %w", err)
 	}
 
 	if _, err := out.WriteString("\n"); err != nil {
-		return "", errors.Wrap(err, "failed to write string")
+		return "", fmt.Errorf("failed to write string: %w", err)
 	}
 
 	return out.String(), nil
@@ -44,19 +41,4 @@ func GetReaderFromFile(tb testing.TB, fPath string) io.ReadCloser {
 	require.NoError(tb, err)
 
 	return file
-}
-
-// GetReportFromFile is a test helper that unmarshal passed filepath into models.Report
-func GetReportFromFile(tb testing.TB, fPath string) models.Report {
-	tb.Helper()
-
-	b, err := ioutil.ReadFile(filepath.Clean(fPath))
-	require.NoError(tb, err)
-
-	var rp models.Report
-
-	err = json.Unmarshal(b, &rp)
-	require.NoError(tb, err)
-
-	return rp
 }
